application comment applicationda qolgan kamchiliklar 1) application added planshet and phone screens resizable devicess

+++ app/src/main/java/uz/idea/mobio/adapters/mainProductAdapter/MainProductChildAdapter.kt
@@ -55,8 +55,7 @@ class MainProductChildAdapter(private val mainProductAdapter: MainProductAdapter
                     product.rate.onEach { rate->
                         rateStar += rate.rating
                     }
-
-                    binding.rate.text =  String.format("%.2f", rateStar/product.rate.size.toDouble())
+                    binding.rate.text = String.format("%.2f", (rateStar/(product.rate.size)))
                 } else {
                     binding.starCard.gone()
                 }
